requirements.txt
#依赖库
#python 版本3.11.2,版本不合适可能导致库的安装解析失败

#发送请求 获取网页内容
requests==2.32.0
#编译js代码 执行js代码
PyExecJS>=1.5.1

#以下是暂时用不到的

#解析网页内容html 提取数据
#beautifulsoup4==4.12.2

#Bs4的解析器 更快更稳定 解析html 提取数据
#lxml==5.0.0

#浏览器自动化库 selenium（用于模拟浏览器登录、验证码等操作）
#selenium==4.15.1

#随机生成user-agent 防止被识别为爬虫 模拟浏览器请求
#fake-useragent==1.4.0


